backend/requirements.txt
Flask
Flask-SQLAlchemy
flask-cors
PyJWT